lib/agents/real-time-collaboration-tools.ts
import { BaseAgent, AgentConfig } from './base';

export interface Real-timeCollaborationToolsConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class Real-timeCollaborationToolsAgent extends BaseAgent {
  private config: Real-timeCollaborationToolsConfig;

  constructor(config: Real-timeCollaborationToolsConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Real-time Collaboration Tools implementation
    await new Promise(resolve => setTimeout(resolve, 1000));
    
    return `Real-time Collaboration Tools completed successfully for: ${input}`;
  }

  async process(data: any): Promise<any> {
    // Advanced real-time collaboration tools processing
    return { status: 'success', data };
  }
}