__tests__/lib/agents/real-time-collaboration-tools.test.ts
import { Real-timeCollaborationToolsAgent } from '@/lib/agents/real-time-collaboration-tools';

describe('Real-time Collaboration ToolsAgent', () => {
  let agent: Real-timeCollaborationToolsAgent;

  beforeEach(() => {
    agent = new Real-timeCollaborationToolsAgent({
      name: 'Test Real-time Collaboration Tools Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      const result = await agent.execute('test input');
      expect(result).toContain('Real-time Collaboration Tools completed successfully');
    });

    it('should handle empty input', async () => {
      const result = await agent.execute('');
      expect(result).toBeDefined();
    });
  });

  describe('process', () => {
    it('should process data correctly', async () => {
      const result = await agent.process({ test: 'data' });
      expect(result.status).toBe('success');
    });
  });
});